Added three personal functionalities. Assignment1 is basically done by now. Need to edit the TODO file though.

+++ ProjectAnalyze.sh
@@ -5,9 +5,8 @@
  
 
 getInformation(){
-	theDate=$(date) && echo "The date today is: $theDate"
+	theDate=$(date) && echo "The time now is: $theDate"
 	theUserName=$(whoami) && echo "Current User is: $theUserName"
-	echo " "
 	return
 }
 
@@ -15,7 +14,7 @@ checkUpdate(){
 	echo "-------------------------------------------------------------"
 	git fetch && git status -uno
 	echo "-------------------------------------------------------------"
-        read -n1 -r -p "Press any key to return to main menu..." key	
+	hold
 }
 
 
@@ -36,16 +35,19 @@ uncommited(){
 	echo "----------------------END OF LOG-----------------------------"
 	
 	fi
-        read -n1 -r -p "Press any key to return to main menu..." key
-
+	hold
 
 }
 
 extractTODO(){
 	echo "-------------------------------------------------------------"
 	echo "Extracting all lines with #TODO in the repository to the todo.log file."
-	rm todo.log
-	grep -rh '#TODO' * > todo.log
+	if [ -e todo.log ]
+	then
+		rm todo.log
+		echo "todo.log already exists. Removing it and creating a new one..."
+	fi
+	grep -rh --exclude={*.log,ProjectAnalyze.sh} '#TODO' *  > todo.log
 	echo "Jobs done."
 	echo "-------------------------------------------------------------"
 	read -p 'Do you want to see the log now?(Y/N)'$'\n' seeLog
@@ -55,8 +57,7 @@ extractTODO(){
 		cat todo.log
 	echo "----------------------END OF LOG-----------------------------"
 	fi
-        read -n1 -r -p "Press any key to return to main menu..." key
-
+	hold
 
 }
 
@@ -75,12 +76,16 @@ findHaskellError(){
 	echo "----------------------END OF LOG-----------------------------"
 	
 	fi
+	hold
+
+}
+hold(){
         read -n1 -r -p "Press any key to return to main menu..." key
-
-
+	
 }
 
 main(){
+	getInformation
 	while true; do
 	read -p $'Please enter a number:\n1.Required Functionalities\n2.Additional Functionalities\n3.Exit\n> ' choice1
 	case "$choice1" in
@@ -91,6 +96,35 @@ main(){
 	esac
 	done
 }
+seeCourse(){
+	read -p $'Please enter a number:\n1.See course schedule.\n2.See grading scale.\n> ' choice3
+        case "$choice3" in
+                1) lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | sed -n '35,57p' && hold;;
+                2) lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | sed -n '58,66p' && hold;;
+		*) ;;
+	esac
+
+}
+
+seeGrade(){
+        echo "-------------------------------------------------------------"
+	L1=$(lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | grep -n '7 Marks' | cut -f1 -d:)
+	L2=$(lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | grep -n '400160537' | cut -f1 -d:)	
+	lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | sed -n "$L1"','"$L2"'p'
+        echo "-------------------------------------------------------------"
+	read -p 'Do you want to export the grades?(Y/N)'$'\n' decision
+	if [[ "$decision" =~ ^(y|yes)$ ]]
+	then
+	if [ -e Grades.txt ]
+	then
+		rm Grades.txt
+		echo "Grades.txt already exists. Removing it and creating a new one..."
+	fi	
+	lynx -dump http://www.cas.mcmaster.ca/~dalvescb/ | sed -n "$L1"','"$L2"'p' > Grades.txt	
+	fi
+	echo "Done."	
+	hold	
+}
 
 subMenu1(){
 	read -p $'Please enter a number:\n1.Check if you repo is up to date.\n2.Ouput uncommited changes.\n3.Find and output all TODOs in the project.\n4.Check and output Haskell errors.\n> ' choice2
@@ -99,13 +133,32 @@ subMenu1(){
                 2) uncommited;;
                 3) extractTODO;;
 		4) findHaskellError;;
-		5) subMenu1;;
-		5) ;;
+		*) subMenu1;;
 	esac
 }
 
-subMenu2(){
-	echo "subMenu2 is also empty"
+backUp(){
+	echo "Starting the back-up process.."
+	cp -R ~/CS1XA3 ~/BackUp
+	tar czvf ~/Backup.tar.gz ~/BackUp
+	rm -rf ~/BackUp
+	echo "Back-up finished. You can now see the file \"Backup.tar.gz\" under your home directory."
+	hold	
 }
+
+starWars(){
+	telnet towel.blinkenlights.nl
+}
+
+subMenu2(){
+	read -p $'Please enter a number:\n1.See course information.\n2.See grades.\n3.Backup the repository.\n> ' choice4
+        case "$choice4" in
+                1) seeCourse;;
+                2) seeGrade;;
+		3) backUp;;
+		4) starWars;;
+		*) subMenu2;;
+	esac
+}	
 
 main
